+++ src/js/components/Fit.js
@@ -1,13 +1,12 @@
 import React, { Component } from "react";
 import { connect } from 'react-redux';
-
+import { localeData } from '../reducers/localization';
 import { getFits, addFit, editFit, removeFit, TOGGLE_FIT_ADD_FORM, TOGGLE_FIT_EDIT_FORM } from '../actions';
 //Components
 import Add from "grommet/components/icons/base/Add";
 import AppHeader from './AppHeader';
 import Box from 'grommet/components/Box';
 import Button from 'grommet/components/Button';
-import Close from "grommet/components/icons/base/Close";
 import Edit from "grommet/components/icons/base/Edit";
 import Footer from 'grommet/components/Footer';
 import Form from 'grommet/components/Form';
@@ -21,21 +20,29 @@ import ListItem from 'grommet/components/ListItem';
 import ListPlaceholder from 'grommet-addons/components/ListPlaceholder';
 import Section from 'grommet/components/Section';
 import Spinning from 'grommet/components/icons/Spinning';
+import Trash from "grommet/components/icons/base/Trash";
 
 class Fit extends Component {
 
   constructor () {
     super();
     this.state = {
-      buyerHref: window.sessionStorage.buyerHref,
+      buyerHref: sessionStorage.buyerHref,
+      buyerName: sessionStorage.buyerName,
+      role: sessionStorage.role,
       fitName: '',
       href: null,
       errors: []
     };
   }
 
-  componentDidMount () {
-    this.props.dispatch(getFits());
+  componentWillMount () {
+    this.setState({localeData: localeData()});
+    const { buyerName } = window.sessionStorage;
+    if (buyerName != 'undefined') {
+      this.props.dispatch(getFits(buyerName));
+    }
+
   }
 
   _addFit () {
@@ -50,7 +57,14 @@ class Fit extends Component {
   }
 
   _removeFit (href) {
-    this.props.dispatch(removeFit(href));
+    if (sessionStorage.privilege == 'USER') {
+      alert('You do not have privilege for the operation.');
+      return;
+    }
+    let value = confirm('Are you sure to delete this Fit?');
+    if (value) {
+      this.props.dispatch(removeFit(href));
+    }
   }
 
   _editFit () {
@@ -65,11 +79,19 @@ class Fit extends Component {
   }
 
   _onEditClick (name, href) {
+    if (sessionStorage.privilege == 'USER') {
+      alert('You do not have privilege for the operation.');
+      return;
+    }
     this.setState({href: href, fitName: name});
     this.props.dispatch({type: TOGGLE_FIT_EDIT_FORM, payload: {editing: true}});
   }
 
   _onAddClick () {
+    if (sessionStorage.privilege == 'USER') {
+      alert('You do not have privilege for the operation.');
+      return;
+    }
     this.props.dispatch({type: TOGGLE_FIT_ADD_FORM, payload: {adding: true}});
   }
 
@@ -88,17 +110,33 @@ class Fit extends Component {
 
 
   render () {
-    //this.props.dispatch(getFits());
+    const {role, buyerName } = window.sessionStorage;
+    const { localeData } = this.state;
+
+    if (role == 'USER' || (role == 'MERCHANT' && buyerName == 'undefined')) {
+      return (
+        <Box>
+  		    <AppHeader page={localeData.label_fit} />
+          <Section>
+            <Box alignSelf="center">
+              <h3>You need to select buyer on Home page becuase you have 'USER' privilege or 'MERCHANT' privilage with no buyer access.</h3>
+            </Box>
+          </Section>
+  			</Box>
+      );
+    }
+
     let { fits, fetching, adding, editing } = this.props.fit;
     let count = fetching ? 100 : fits.length;
     let items = fits.map(fit => {
       return (
         <ListItem key={fit.href} justify="between" pad={{vertical:'none',horizontal:'small'}} >
           <span> {fit.name} </span>
-          <span className="secondary">
-            <Button icon={<Edit />} onClick={this._onEditClick.bind(this, fit.name, fit.href)}/>
-            <Button icon={<Close />} onClick={this._removeFit.bind(this, fit.href)} />
-          </span>
+            <span className="secondary">
+              <Button icon={<Edit />} onClick={this._onEditClick.bind(this, fit.name, fit.href)} />
+              <Button icon={<Trash />} onClick={this._removeFit.bind(this, fit.href)} />
+            </span>
+
         </ListItem>
       );
     });
@@ -145,7 +183,7 @@ class Fit extends Component {
               {loading}
             </Box>
             <Box size="large" alignSelf="center" >
-              <List selectable={true} > {items} </List>
+              <List > {items} </List>
               <ListPlaceholder unfilteredTotal={count} filteredTotal={count} emptyMessage="You do not have any fits at the moment." />
             </Box>
             <Box size="small" alignSelf="center" pad={{vertical:'large'}}>
@@ -161,7 +199,7 @@ class Fit extends Component {
 }
 
 let select = (store) => {
-  return { fit: store.fit};
+  return { fit: store.fit, user: store.user};
 };
 
 export default connect(select)(Fit);
